Clear the right uid's alarms, please
When we clear data for an app, we need to cancel the alarms for
that app's uid -- not the *calling* uid, which could be something
like the Settings app's, i.e. the system uid.  That's bad.

Bug: 72059961
Bug: 70670726
Test: manual
Change-Id: Ib6833ce2672bc6dfab6a4be6014a0301611576e2
(cherry picked from commit 60bcff89f670c76f9fc6bb26682c1c8f68e099f7)