Remove @hide from incremental flags & return code.
See ag/3457616 for the implementation of the flags.
See ag/3498077 for the implementation of the return code

These changes are related because the transport uses the value of the
incremental/non-incremental flag to decide whether to request a
non-incremental backup using the new return code.

I chose to add both an INCREMENTAL and NON_INCREMENTAL flag to allow the
transport to distinguish between both cases and older versions of the OS
not supporting this flag.

Bug: 71792427
Change-Id: Ifa099fb73de70eb3b5b315cc3b266c308dfb719d
Test: None required, only changing javadoc.